requirements.txt
streamlit
PyMuPDF
Pillow
pyzbar